bump kool from 0.1.28 to 0.1.35, fix test with breaking change

+++ src/test/java/com/github/davidmoten/aws/lw/client/RuntimeAnalysisTest.java
@@ -131,7 +131,7 @@ public class RuntimeAnalysisTest {
                                 .mapWithIndex() //
                                 .groupByList( //
                                         HashMap::new, //
-                                        y -> y.index() % 3, //
+                                        y -> (int) (y.index() % 3), //
                                         y -> y.value())
                                 .get());
